requirements.txt
# Core packages
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0

# NLP Libraries
nltk>=3.7.0
spacy>=3.4.0
gensim>=4.2.0

# Visualization
pyLDAvis>=3.3.0
wordcloud>=1.9.0

# Testing
pytest>=7.0.0

# Additional utilities
tqdm>=4.64.0
scipy>=1.7.0
PyYAML>=6.0